+++ src/components/MinMax.js
