+++ src/components/MinMaxFn.js
