+++ src/components/MiniMaxInput.js
