+++ src/components/MiniMaxInputFn.js
